Separated events form dialog from data table
